Replace custom authenticator passport with custom badge usage
